Fix errors in drawing of time scales
Fix errors in drawing of time scales on both images of summary  index tracks and false-colour spectrogrms.
there were multiple problems. First problem was in AnalyseLongRecording.cs where a fixed timescale was cculated. This now changed so that the correct IndexcalculationDuration value is passed to drawing methods.
Needed to edit Sandpit.cs to test these fixes.

Changed default setting for second ColourMap in the Config file.